Separate directory for entrypoints
