+++ src/ggrc/assets/javascripts/entrypoints/my-assessments/index.js
@@ -1,0 +1,56 @@
+/*!
+ Copyright (C) 2017 Google Inc.
+ Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+ */
+
+import '../../permission';
+import '../../bootstrap/sticky-popover';
+import '../../bootstrap/modal-ajax';
+import '../../bootstrap/modal-form';
+import '../../plugins/lodash_helpers';
+import '../../plugins/persistent_notifier';
+import '../../ggrc_base';
+import '../../apps/base_widgets';
+import '../../application';
+import '../../plugins/ggrc_utils';
+import '../../plugins/utils/advanced-search-utils';
+import '../../plugins/utils/ca-utils';
+import '../../plugins/utils/query-api-utils';
+import '../../plugins/utils/current-page-utils';
+import '../../plugins/utils/state-utils';
+import '../../plugins/utils/ggrcq-utils';
+import '../../plugins/utils/tree-view-utils';
+import '../../plugins/utils/dashboards-utils';
+import '../../plugins/datepicker';
+import '../../plugins/can_control';
+import '../../plugins/autocomplete';
+import '../../plugins/ajax_extensions';
+import '../../plugins/canjs_extensions';
+import '../../plugins/component_registry';
+import '../../plugins/openclose';
+import '../../plugins/tooltip';
+import '../../plugins/popover';
+import '../../plugins/popover_template';
+import '../../mustache_helper';
+import '../../generated/ggrc_filter_query_parser';
+
+// Models
+import '../../models';
+
+// Controllers
+import '../../controllers/dashboard_widgets_controller';
+import '../../controllers/modals_controller';
+import '../../controllers/delete_modal_controller';
+import '../../controllers/unmap_modal_controller';
+import '../../controllers/info_pin_controller';
+import '../../controllers/automapper_controller';
+import '../../controllers/dashboard_controller';
+import '../../controllers/quick_form_controller';
+import '../../controllers/mapper';
+
+// Modules and Apps
+import '../../modules/widget_descriptor';
+import '../../modules/widget_list';
+
+import '../../apps/business_objects';
+import '../../apps/custom_attributes';
