+++ src/ggrc/assets/javascripts/entrypoints/workflow/index.js
@@ -1,0 +1,4 @@
+/*!
+    Copyright (C) 2017 Google Inc.
+    Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
+*/
