--- src/ggrc/assets/javascripts/import.js
@@ -1,6 +1,0 @@
-/*!
-   Copyright (C) 2017 Google Inc.
-   Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-import './components/csv/import';
