--- src/ggrc/assets/javascripts/my-assessments.js
@@ -1,56 +1,0 @@
-/*!
-  Copyright (C) 2017 Google Inc.
-  Licensed under http://www.apache.org/licenses/LICENSE-2.0 <see LICENSE file>
-*/
-
-import './permission.js';
-import './bootstrap/sticky-popover.js';
-import './bootstrap/modal-ajax.js';
-import './bootstrap/modal-form.js';
-import './plugins/lodash_helpers.js';
-import './plugins/persistent_notifier.js';
-import './ggrc_base.js';
-import './apps/base_widgets.js';
-import './application.js';
-import './plugins/ggrc_utils.js';
-import './plugins/utils/advanced-search-utils.js';
-import './plugins/utils/ca-utils.js';
-import './plugins/utils/query-api-utils.js';
-import './plugins/utils/current-page-utils.js';
-import './plugins/utils/state-utils.js';
-import './plugins/utils/ggrcq-utils.js';
-import './plugins/utils/tree-view-utils.js';
-import './plugins/utils/dashboards-utils.js';
-import './plugins/datepicker.js';
-import './plugins/can_control.js';
-import './plugins/autocomplete.js';
-import './plugins/ajax_extensions.js';
-import './plugins/canjs_extensions.js';
-import './plugins/component_registry.js';
-import './plugins/openclose.js';
-import './plugins/tooltip.js';
-import './plugins/popover.js';
-import './plugins/popover_template.js';
-import './mustache_helper.js';
-import './generated/ggrc_filter_query_parser.js';
-
-// Models
-import './models';
-
-// Controllers
-import './controllers/dashboard_widgets_controller';
-import './controllers/modals_controller';
-import './controllers/delete_modal_controller';
-import './controllers/unmap_modal_controller';
-import './controllers/info_pin_controller';
-import './controllers/automapper_controller';
-import './controllers/dashboard_controller';
-import './controllers/quick_form_controller';
-import './controllers/mapper';
-
-// Modules and Apps
-import './modules/widget_descriptor';
-import './modules/widget_list';
-
-import './apps/business_objects';
-import './apps/custom_attributes';
